requirements.txt
pandas
numpy
openpyxl
xlsxwriter
scipy
matplotlib
seaborn
requests
python-dotenv
rapidfuzz